Added WebGL 'length' expressions
The expression observes the syntax ['length', <value>] where <value> is a string, an array of numbers, or an 'array' expression. The 'length' expression evaluates to a number giving the length of the string if <value> is a string, the length of the array if <value> is a number array, or the number of numbers in the resulting vector if <value> is an 'array' expression.

Added unit tests to check proper expression evaluation and error handling.